Added Sort to Save Cards to Stop Deletion Bug
Added Sort to Save Cards in order to stop deletion bug.

how the bug worked was that the save button would put the card in position card+1 when a card was removed the cards would not change position meaning a card could be at the spot card+1 so the sort makes them change position.

positions:

1     2
3     4
5     6
7     8
ect...